refactor: Ticket presentation model 파일로 이동
