+++ app/src/main/java/woowacourse/movie/presentation/ui/detail/DetailUiModel.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.presentation.ui.detail
 
 import woowacourse.movie.domain.model.Screen
 import woowacourse.movie.domain.model.ScreenDate
-import woowacourse.movie.domain.model.Ticket
+import woowacourse.movie.presentation.model.Ticket
 import java.time.LocalTime
 
 data class DetailUiModel(
